+++ src/main.cpp
@@ -224,7 +224,7 @@ int main(int argc, char *argv[]) {
 	if (argc > 1) {
 		int len, i;
 		for (len = 1, i = 1; i < argc; i++) {
-			len += strlen(argv[i]) + 1;
+			len += (int)strlen(argv[i]) + 1;
 		}
 
 		char *cmdline = (char *)malloc(len);
